Move change due date API tests to be together CIRC-647
